# Performance test: String flag vs edge flag