--- a/graph/api/src/main/java/org/jboss/windup/graph/model/resource/ApplicationFlag.java
+++ b/graph/api/src/main/java/org/jboss/windup/graph/model/resource/ApplicationFlag.java
@@ -1,0 +1,6 @@
+package org.jboss.windup.graph.model.resource;
+
+public enum ApplicationFlag
+{
+    APPLICATION,SERVER
+}