--- a/graph/api/src/main/java/org/jboss/windup/graph/model/resource/ApplicationFlagVertex.java
+++ b/graph/api/src/main/java/org/jboss/windup/graph/model/resource/ApplicationFlagVertex.java
@@ -1,0 +1,11 @@
+package org.jboss.windup.graph.model.resource;
+
+import org.jboss.windup.graph.model.WindupVertexFrame;
+
+import com.tinkerpop.frames.modules.typedgraph.TypeValue;
+
+@TypeValue("ApplicationFlagVertex")
+public interface ApplicationFlagVertex extends WindupVertexFrame
+{
+    
+}