--- a/tests/src/test/java/org/jboss/windup/tests/application/WindupArchitectureMediumBinaryModeTest.java
+++ b/tests/src/test/java/org/jboss/windup/tests/application/WindupArchitectureMediumBinaryModeTest.java
@@ -1,27 +1,42 @@
 package org.jboss.windup.tests.application;
 
 import java.io.File;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.io.FileWriter;
+import java.io.PrintWriter;
+import java.util.logging.Logger;
+
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.forge.arquillian.AddonDependencies;
 import org.jboss.forge.arquillian.AddonDependency;
 import org.jboss.forge.arquillian.archive.AddonArchive;
+import org.jboss.forge.furnace.repositories.AddonDependencyEntry;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.windup.config.AbstractRuleProvider;
+import org.jboss.windup.config.GraphRewrite;
+import org.jboss.windup.config.RuleSubset;
+import org.jboss.windup.config.metadata.MetadataBuilder;
+import org.jboss.windup.config.operation.GraphOperation;
+import org.jboss.windup.config.operation.Iteration;
+import org.jboss.windup.config.operation.iteration.AbstractIterationOperation;
+import org.jboss.windup.config.query.Query;
+import org.jboss.windup.tests.application.WindupArchitectureTest;
+import org.jboss.windup.config.query.QueryGremlinCriterion;
 import org.jboss.windup.graph.GraphContext;
-import org.jboss.windup.reporting.model.ReportModel;
-import org.jboss.windup.reporting.service.ReportService;
-import org.jboss.windup.rules.apps.java.ip.CreateStaticIPAddressReportRuleProvider;
-import org.jboss.windup.rules.apps.java.model.JarManifestModel;
-import org.jboss.windup.rules.apps.java.reporting.rules.CreateJavaApplicationOverviewReportRuleProvider;
-import org.jboss.windup.rules.apps.java.service.JarManifestService;
-import org.jboss.windup.testutil.html.TestJavaApplicationOverviewUtil;
-import org.jboss.windup.testutil.html.TestStaticIPReportUtil;
-import org.junit.Assert;
+import org.jboss.windup.graph.model.resource.ApplicationFlag;
+import org.jboss.windup.graph.model.resource.FileModel;
+import org.jboss.windup.rules.apps.java.scan.provider.AnalyzeJavaFilesRuleProvider;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.ocpsoft.rewrite.config.Configuration;
+import org.ocpsoft.rewrite.config.ConfigurationBuilder;
+import org.ocpsoft.rewrite.context.EvaluationContext;
+
+import com.tinkerpop.blueprints.Vertex;
+import com.tinkerpop.gremlin.java.GremlinPipeline;
 
 @RunWith(Arquillian.class)
 public class WindupArchitectureMediumBinaryModeTest extends WindupArchitectureTest
@@ -34,19 +49,32 @@ public class WindupArchitectureMediumBinaryModeTest extends WindupArchitectureTe
                 @AddonDependency(name = "org.jboss.windup.exec:windup-exec"),
                 @AddonDependency(name = "org.jboss.windup.rules.apps:windup-rules-java"),
                 @AddonDependency(name = "org.jboss.windup.rules.apps:windup-rules-java-ee"),
+                @AddonDependency(name = "org.jboss.forge.furnace.container:cdi"),
                 @AddonDependency(name = "org.jboss.windup.rules.apps:windup-rules-tattletale"),
                 @AddonDependency(name = "org.jboss.windup.tests:test-util"),
                 @AddonDependency(name = "org.jboss.windup.config:windup-config-groovy"),
-                @AddonDependency(name = "org.jboss.forge.furnace.container:cdi"),
     })
     public static AddonArchive getDeployment()
     {
         return ShrinkWrap.create(AddonArchive.class)
                     .addBeansXML()
-                    .addClass(WindupArchitectureTest.class)
-                    .addAsResource(new File("src/test/groovy/GroovyExampleRule.windup.groovy"));
+                    .addClasses(WindupArchitectureTest.class,WindupArchitectureMediumBinaryModeTest.class)
+                    .addAsAddonDependencies(
+                                AddonDependencyEntry.create("org.jboss.windup.graph:windup-graph"),
+                                AddonDependencyEntry.create("org.jboss.windup.reporting:windup-reporting"),
+                                AddonDependencyEntry.create("org.jboss.windup.exec:windup-exec"),
+                                AddonDependencyEntry.create("org.jboss.windup.rules.apps:windup-rules-java"),
+                                AddonDependencyEntry.create("org.jboss.windup.rules.apps:windup-rules-java-ee"),
+                                AddonDependencyEntry.create("org.jboss.windup.rules.apps:windup-rules-tattletale"),
+                                AddonDependencyEntry.create("org.jboss.windup.tests:test-util"),
+                                AddonDependencyEntry.create("org.jboss.forge.furnace.container:cdi"),
+                                AddonDependencyEntry.create("org.jboss.windup.config:windup-config-groovy")
+                    );
     }
 
+    @Inject
+    JavaClassTestRuleProvider rp;
+    File resultFile = new File("/home/mbriskar/performance.txt");
     @Test
     public void testRunWindupMedium() throws Exception
     {
@@ -55,75 +83,205 @@ public class WindupArchitectureMediumBinaryModeTest extends WindupArchitectureTe
         try (GraphContext context = createGraphContext())
         {
             super.runTest(context, path, false);
-            validateManifestEntries(context);
-            validateReports(context);
+            System.out.println(rp.firstRuleMatchCount);
+            PrintWriter writer = new PrintWriter(new FileWriter(resultFile,true));
+            writer.println();
+            writer.append("The first rule using String flag results:");
+            writer.println();
+            writer.append("String flag match count: " + rp.firstRuleMatchCount);
+            writer.println();
+            writer.append("String flag time: " + rp.firstExecution);
+            writer.println();
+            writer.append("Edge flag match count: " + rp.secondRuleMatchCount);
+            writer.println();
+            writer.append("Edge flag time: " + rp.secondExecution);
+            writer.println();
+            writer.close();
+            System.out.println(rp.secondRuleMatchCount);
+            System.out.println("finish");
         }
     }
-
-    private void validateManifestEntries(GraphContext context) throws Exception
+    
+    @Test
+    public void testRunWindupMedium2() throws Exception
     {
-        JarManifestService jarManifestService = new JarManifestService(context);
-        Iterable<JarManifestModel> manifests = jarManifestService.findAll();
+        final String path = "../test-files/Windup1x-javaee-example.war";
 
-        int numberFound = 0;
-        boolean warManifestFound = false;
-        for (JarManifestModel manifest : manifests)
+        try (GraphContext context = createGraphContext())
         {
-            if (manifest.getArchive().getFileName().equals("Windup1x-javaee-example.war"))
-            {
-                Assert.assertEquals("1.0", manifest.asVertex().getProperty("Manifest-Version"));
-                Assert.assertEquals("Plexus Archiver", manifest.asVertex().getProperty("Archiver-Version"));
-                Assert.assertEquals("Apache Maven", manifest.asVertex().getProperty("Created-By"));
-                warManifestFound = true;
-            }
-
-            numberFound++;
+            super.runTest(context, path, false);
+            System.out.println(rp.firstRuleMatchCount);
+            PrintWriter writer = new PrintWriter(new FileWriter(resultFile,true));
+            writer.println();
+            writer.append("The first rule using String flag results:");
+            writer.println();
+            writer.append("String flag match count: " + rp.firstRuleMatchCount);
+            writer.println();
+            writer.append("String flag time: " + rp.firstExecution);
+            writer.println();
+            writer.append("Edge flag match count: " + rp.secondRuleMatchCount);
+            writer.println();
+            writer.append("Edge flag time: " + rp.secondExecution);
+            writer.println();
+            writer.close();
+            System.out.println(rp.secondRuleMatchCount);
+            System.out.println("finish");
         }
-        Assert.assertEquals(9, numberFound);
-        Assert.assertTrue(warManifestFound);
     }
-
-    private void validateStaticIPReport(GraphContext context)
+    
+    @Test
+    public void testRunWindupMedium3() throws Exception
     {
-        ReportService reportService = new ReportService(context);
-        ReportModel reportModel = reportService.getUniqueByProperty(
-                    ReportModel.TEMPLATE_PATH,
-                    CreateStaticIPAddressReportRuleProvider.TEMPLATE_REPORT);
-        TestStaticIPReportUtil util = new TestStaticIPReportUtil();
-        Path reportPath = Paths.get(reportService.getReportDirectory(), reportModel.getReportFilename());
-        util.loadPage(reportPath);
-        Assert.assertTrue(util
-                    .checkStaticIPInReport(
-                                "org.apache.wicket.protocol.http.mock.MockHttpServletRequest",
-                                "Line Number 80, Column Number 25", "127.0.0.1"));
-        Assert.assertTrue(util
-                    .checkStaticIPInReport(
-                                "org.apache.wicket.protocol.http.mock.MockHttpServletRequest",
-                                "Line Number 613, Column Number 16", "127.0.0.1"));
-        Assert.assertTrue(util
-                    .checkStaticIPInReport(
-                                "org.apache.wicket.protocol.http.mock.MockHttpServletRequest",
-                                "Line Number 616, Column Number 16", "127.0.0.1"));
+        final String path = "../test-files/Windup1x-javaee-example.war";
+
+        try (GraphContext context = createGraphContext())
+        {
+            super.runTest(context, path, false);
+            System.out.println(rp.firstRuleMatchCount);
+            PrintWriter writer = new PrintWriter(new FileWriter(resultFile,true));
+            writer.println();
+            writer.append("The first rule using String flag results:");
+            writer.println();
+            writer.append("String flag match count: " + rp.firstRuleMatchCount);
+            writer.println();
+            writer.append("String flag time: " + rp.firstExecution);
+            writer.println();
+            writer.append("Edge flag match count: " + rp.secondRuleMatchCount);
+            writer.println();
+            writer.append("Edge flag time: " + rp.secondExecution);
+            writer.println();
+            writer.close();
+            System.out.println(rp.secondRuleMatchCount);
+            System.out.println("finish");
+        }
     }
 
-    /**
-     * Validate that the report pages were generated correctly
-     */
-    private void validateReports(GraphContext context)
+    @Singleton
+    public static class JavaClassTestRuleProvider extends AbstractRuleProvider
     {
-        ReportService reportService = new ReportService(context);
-        ReportModel reportModel = reportService.getUniqueByProperty(
-                    ReportModel.TEMPLATE_PATH,
-                    CreateJavaApplicationOverviewReportRuleProvider.TEMPLATE_APPLICATION_REPORT);
-        Path appReportPath = Paths.get(reportService.getReportDirectory(), reportModel.getReportFilename());
+        private static Logger log = Logger.getLogger(RuleSubset.class.getName());
 
-        TestJavaApplicationOverviewUtil util = new TestJavaApplicationOverviewUtil();
-        util.loadPage(appReportPath);
-        util.checkFilePathAndTag("Windup1x-javaee-example.war", "META-INF/maven/javaee/javaee/pom.properties",
-                    "Properties");
-        util.checkFilePathEffort("Windup1x-javaee-example.war", "META-INF/maven/javaee/javaee/pom.properties", 0);
-        util.checkFilePathEffort("Windup1x-javaee-example.war/WEB-INF/lib/joda-time-2.0.jar",
-                    "org.joda.time.tz.DateTimeZoneBuilder", 32);
-        validateStaticIPReport(context);
+        private int firstRuleMatchCount = 0;
+        private long firstExecution = 0;
+
+        private int secondRuleMatchCount = 0;
+        private long secondExecution = 0;
+
+        public JavaClassTestRuleProvider()
+        {
+            super(MetadataBuilder.forProvider(JavaClassTestRuleProvider.class)
+                        .addExecuteAfter(AnalyzeJavaFilesRuleProvider.class));
+        }
+
+        // @formatter:off
+        @Override
+        public Configuration getConfiguration(GraphContext context)
+        {
+            return ConfigurationBuilder.begin()
+                        .addRule().perform(new GraphOperation()
+                            {
+
+                                @Override
+                                public void perform(GraphRewrite event, EvaluationContext context)
+                                {
+                                    firstExecution=System.currentTimeMillis();
+                                }
+                                
+                            }
+                        )
+            .addRule().when(
+                        Query.fromType(FileModel.class).withProperty(FileModel.APPLICATION_FLAG, "application")
+            ).perform(
+                Iteration.over().perform(new AbstractIterationOperation<FileModel>()
+                {
+                    @Override
+                    public void perform(GraphRewrite event, EvaluationContext context, FileModel payload)
+                    {
+                        firstRuleMatchCount++;
+                    }
+                }).endIteration()
+            )
+            .addRule().perform(
+                            new GraphOperation()
+                            {
+
+                                @Override
+                                public void perform(GraphRewrite event, EvaluationContext context)
+                                {
+                                    firstExecution=System.currentTimeMillis() - firstExecution;
+                                }
+                                
+                            }
+                        )
+                        
+                        
+                        
+                        
+                        
+                        
+                        
+                        //now the edge
+                        
+                        
+                        
+                        
+                        .addRule().perform(
+                            new GraphOperation()
+                            {
+
+                                @Override
+                                public void perform(GraphRewrite event, EvaluationContext context)
+                                {
+                                    secondExecution=System.currentTimeMillis();
+                                }
+                                
+                            }
+                        )
+            .addRule().when(
+                        Query.fromType(FileModel.class).piped(new QueryGremlinCriterion() {
+
+                            @Override
+                            public void query(GraphRewrite event, GremlinPipeline<Vertex, Vertex> pipeline)
+                            {
+                                pipeline.as("result").out("ApplicationFlagVertex").back("result");
+                            }
+                            
+                        })
+            ).perform(
+                Iteration.over().perform(new AbstractIterationOperation<FileModel>()
+                {
+                    @Override
+                    public void perform(GraphRewrite event, EvaluationContext context, FileModel payload)
+                    {
+                        secondRuleMatchCount++;
+                    }
+                }).endIteration()
+            )
+            .addRule().perform(
+                           new GraphOperation()
+                            {
+
+                                @Override
+                                public void perform(GraphRewrite event, EvaluationContext context)
+                                {
+                                    secondExecution=System.currentTimeMillis() - secondExecution;
+                                }
+                                
+                            })
+                        ;
+        }
+        // @formatter:on
+
+        public int getFirstRuleMatchCount()
+        {
+            return firstRuleMatchCount;
+        }
+
+        public int getSecondRuleMatchCount()
+        {
+            return secondRuleMatchCount;
+        }
+
     }
+
 }