PLEASE DO NOT MERGE THIS. It is done just to provide test for string vs. edge performance test as fast as possible. According to the output, string property is faster than edge:

1st run
String flag match count: 3223 (number of vertices, is not cleared between different @Test methods. Just to check both matched the same ammount of vertices)
String flag time: 443 ms (time of the rule execution)
Edge flag match count: 3223
Edge flag time: 421 ms

String flag match count: 6446
String flag time: 300 ms
Edge flag match count: 6446
Edge flag time: 443 ms


String flag match count: 9669
String flag time: 323 ms
Edge flag match count: 9669
Edge flag time: 423 ms

2nd run
String flag match count: 3223
String flag time: 289 ms
Edge flag match count: 3223
Edge flag time: 489 ms

String flag match count: 6446
String flag time: 309 ms
Edge flag match count: 6446
Edge flag time: 650 ms


String flag match count: 9669
String flag time: 237 ms
Edge flag match count: 9669
Edge flag time: 486 ms

 * Even after using gremlin for property (``` pipeline.has() ``` ) and removing the @Index, the result is almost the same. Only the first query is slower for the property, but since then the string property match is faster in this query . 

 * Using ``` pipeline.as("result").out("ApplicationFlagVertex").back("result") ``` without fromType() (that results in starting from all vertices) is a disaster with 10-20x bigger time