@windupgithubbot1 2015-04-30
Refer to this link for build results (access rights to CI server needed): 
http://10.3.8.5//job/windup-pr-builder/708/<h2>Failed Tests: <span class='status-failure'>5</span></h2><h3><a name='org.jboss.windup.tests:windup-tests' /><a href='http://10.3.8.5/job/windup-pr-builder/708/org.jboss.windup.tests$windup-tests/testReport'>org.jboss.windup.tests:windup-tests</a>: <span class='status-failure'>5</span></h3><ul><li><a href='http://10.3.8.5/job/windup-pr-builder/708/org.jboss.windup.tests$windup-tests/testReport/org.jboss.windup.tests.application/WindupArchitectureMediumBinaryModeTest/testRunWindupMedium2/'><strong>org.jboss.windup.tests.application.WindupArchitectureMediumBinaryModeTest.testRunWindupMedium2</strong></a></li><li><a href='http://10.3.8.5/job/windup-pr-builder/708/org.jboss.windup.tests$windup-tests/testReport/org.jboss.windup.tests.application/WindupArchitectureMediumBinaryModeTest/testRunWindupMedium3/'><strong>org.jboss.windup.tests.application.WindupArchitectureMediumBinaryModeTest.testRunWindupMedium3</strong></a></li><li><a href='http://10.3.8.5/job/windup-pr-builder/708/org.jboss.windup.tests$windup-tests/testReport/org.jboss.windup.tests.application/WindupArchitectureMediumBinaryModeTest/testRunWindupMedium/'><strong>org.jboss.windup.tests.application.WindupArchitectureMediumBinaryModeTest.testRunWindupMedium</strong></a></li><li><a href='http://10.3.8.5/job/windup-pr-builder/708/org.jboss.windup.tests$windup-tests/testReport/org.jboss.windup.tests.application/WindupArchitectureScanPackagesTest/testRunWindupScanPackages/'><strong>org.jboss.windup.tests.application.WindupArchitectureScanPackagesTest.testRunWindupScanPackages</strong></a></li><li><a href='http://10.3.8.5/job/windup-pr-builder/708/org.jboss.windup.tests$windup-tests/testReport/org.jboss.windup.tests.application/WindupArchitectureSmallBinaryModeTest/testRunWindupTiny/'><strong>org.jboss.windup.tests.application.WindupArchitectureSmallBinaryModeTest.testRunWindupTiny</strong></a></li></ul>
Test FAILed.