@windupgithubbot1 2015-04-30
Refer to this link for build results (access rights to CI server needed): 
http://10.3.8.5//job/windup-core-PR-dependents-check/80/
Test PASSed.